refactor(grafana-ui): move getQueryFieldConfig out of CodeMirror dir
