+++ packages/grafana-ui/src/unstable.ts
@@ -29,7 +29,7 @@ export type {
   SignatureInformation,
   SignatureParameter,
 } from './components/CodeMirror/types';
-export { getQueryFieldConfig } from './components/CodeMirror/queryField';
-export type { QueryFieldConfig, QueryFieldConfigOptions } from './components/CodeMirror/queryField';
+export { getQueryFieldConfig } from './components/QueryFieldConfig/queryFieldConfig';
+export type { QueryFieldConfig, QueryFieldConfigOptions } from './components/QueryFieldConfig/queryFieldConfig';
 export type { SignatureHelpOptions } from './components/CodeMirror/signatureHelp';
 export { TableNG } from './components/Table/TableNG/TableNG';
